refactor app organization, tabs to spaces :)
